feat(setup): minikube dashboard (removed the kubernetes dashboard)

+++ setup/setup__minikube.bash
@@ -1,10 +1,19 @@
 #!/bin/bash
 
-curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube_latest_amd64.deb
-sudo dpkg -i minikube_latest_amd64.deb
-rm minikube_latest_amd64.deb
+if [[ ! -f "/usr/bin/minikube" ]]; then
+    curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube_latest_amd64.deb
+    sudo dpkg -i minikube_latest_amd64.deb
+    rm minikube_latest_amd64.deb
+fi
 
-eval $(ssh-agent -s)
-ssh-add
+unset KUBECONFIG
+if [ $(minikube status | grep Running | wc -l) -gt 0 ]; then
+    echo "minikube is already started"
+else
+    eval $(ssh-agent -s)
+    ssh-add
 
-minikube start --addons registry --addons metrics-server --driver=docker --mount-string "$SSH_AUTH_SOCK:/ssh-agent" --cpus=$HOME_MINIKUBE_CPU --memory=$HOME_MINIKUBE_MEM
+    minikube config set cpus $HOME_MINIKUBE_CPU
+    minikube config set memory $HOME_MINIKUBE_MEM
+    minikube start --addons registry --addons metrics-server --driver=docker --mount-string "$SSH_AUTH_SOCK:/ssh-agent"
+fi
